data_modifier/source/requirements.txt
pandas
sqlalchemy
python-dotenv
psycopg2
numpy